requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
dj-database-url==2.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==23.2
django-graphql-jwt==0.3.4
djangorestframework==3.14.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.4.0
promise==2.3
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.6.3
whitenoise==6.4.0
zipp==3.15.0
